perf: Port TargetAdaptor
